Hash functions implementation move to frost package
The idea is to keep the FROST-algorithm-related components in the frost
package and later, have the roast package import from frost.

This should allow us to better organize the code and reduce developer's
cognitive effort.
