+++ frost/hash_test.go
@@ -1,4 +1,4 @@
-package roast
+package frost
 
 import (
 	"bytes"
